percents

+++ app/src/main/java/com/example/myapplicationtest/Query.java
@@ -43,8 +43,8 @@ public class Query {
                 "WHERE artists.artist_id IN ";
         // check that this features are not null
         String notNull=" AND songs.song_loudness IS NOT NULL AND songs.song_tempo IS NOT NULL";
-        float num_tempo;
-        float num_loudness;
+        double num_tempo;
+        double num_loudness;
         switch (tempo){
             case "Slow":
                 temp=1;
@@ -57,64 +57,64 @@ public class Query {
                 break;
         }
         // according to the loudness and tempo chosen by the user creates the continuation of the query.
-        float numLoud[] = Maps.getInstance().PutInloudness(loudness);
-        float numTempo[] = Maps.getInstance().PutInTempo(tempo);
+        double numLoud[] = Maps.getInstance().PutInloudness(loudness);
+        double numTempo[] = Maps.getInstance().PutInTempo(tempo);
         switch (loudness){
             case "Weak":
-                num_loudness = numLoud[0] - (float)priority.get(prioLoudness);
+                num_loudness = numLoud[0] - (double)priority.get(prioLoudness);
                 if(temp==1){
-                    num_tempo = numTempo[0] + (float)priority.get(prioTempo);
+                    num_tempo = numTempo[0] + (double)priority.get(prioTempo);
                     q=  mapGenre+ "(Select song_artist_id FROM songs where song_tempo <\""+num_tempo+"\""+ "AND song_loudness >\""+num_loudness+"\""+
                             notNull+")";
                 }else if(temp==2){
-                    num_tempo = numTempo[0] - ((float)priority.get(prioTempo)/2);
-                    float num_tempo2 = numTempo[1] + ((float)priority.get(prioTempo)/2);
+                    num_tempo = numTempo[0] - ((double)priority.get(prioTempo)/2);
+                    double num_tempo2 = numTempo[1] + ((double)priority.get(prioTempo)/2);
                     q= mapGenre+
                             "(Select song_artist_id FROM songs WHERE song_tempo between \""+num_tempo+"\""+ "and \""+num_tempo2+"\"" + "AND song_loudness>\""+num_loudness+"\""+
                             notNull+")";
                 }else {
-                    num_tempo = numTempo[0] - (float) priority.get(prioTempo);
+                    num_tempo = numTempo[0] - (double) priority.get(prioTempo);
                     q = mapGenre +
                             "(Select song_artist_id FROM songs WHERE song_tempo>\"" + num_tempo + "\"" + "AND song_loudness>\"" + num_loudness + "\"" +
                             notNull + ")";
                 }
                 break;
             case "Normal":
-                num_loudness = (float)numLoud[0] - ((float)priority.get(prioLoudness)/2);
-                float num_loudness2 = (float)numLoud[1] + ((float)priority.get(prioLoudness)/2);
+                num_loudness = (double)numLoud[0] - ((double)priority.get(prioLoudness)/2);
+                double num_loudness2 = (double)numLoud[1] + ((double)priority.get(prioLoudness)/2);
                 if(temp==1){
-                    num_tempo = numTempo[0] + (float)priority.get(prioTempo);
+                    num_tempo = numTempo[0] + (double)priority.get(prioTempo);
                     q= mapGenre+
                             "(Select song_artist_id FROM songs WHERE song_tempo<\""+num_tempo+"\"" + " AND song_loudness BETWEEN \""+num_loudness+"\"" + "and \""+num_loudness2+"\""+
                             notNull+")";
                 }else if(temp==2){
-                    num_tempo = numTempo[0] - ((float)priority.get(prioTempo)/2);
-                    float num_tempo2 = numTempo[1] + ((float)priority.get(prioTempo)/2);
+                    num_tempo = numTempo[0] - ((double)priority.get(prioTempo)/2);
+                    double num_tempo2 = numTempo[1] + ((double)priority.get(prioTempo)/2);
                     q= mapGenre+
                             "(Select song_artist_id FROM songs WHERE song_tempo between \""+num_tempo+"\"" + "and \""+num_tempo2+"\"" + "AND song_loudness BETWEEN \""+num_loudness+"\"" + "and \""+num_loudness2+"\"" +
                             notNull+")";
                 }else {
-                    num_tempo = numTempo[0] - (float) priority.get(prioTempo);
+                    num_tempo = numTempo[0] - (double) priority.get(prioTempo);
                     q = mapGenre +
                             "(Select song_artist_id FROM songs WHERE song_tempo>\"" + num_tempo + "\"" + "AND song_loudness BETWEEN \"" + num_loudness + "\"" + "and \"" + num_loudness2 + "\"" +
                             notNull + ")";
                 }
                 break;
             case "Strong":
-                num_loudness = numLoud[0] + (float)priority.get(prioLoudness);
+                num_loudness = numLoud[0] + (double)priority.get(prioLoudness);
                 if(temp==1){
-                    num_tempo = numTempo[0] + (float)priority.get(prioTempo);
+                    num_tempo = numTempo[0] + (double)priority.get(prioTempo);
                     q= mapGenre+
                             "(Select song_artist_id FROM songs WHERE song_tempo<\""+num_tempo+"\"" + " AND song_loudness<\""+num_loudness+"\"" +
                             notNull+")";
                 }else if(temp==2){
-                    num_tempo = numTempo[0] - ((float)priority.get(prioTempo)/2);
-                    float num_tempo2 = numTempo[1] + ((float)priority.get(prioTempo)/2);
+                    num_tempo = numTempo[0] - ((double)priority.get(prioTempo)/2);
+                    double num_tempo2 = numTempo[1] + ((double)priority.get(prioTempo)/2);
                     q= mapGenre+
                             "(Select song_artist_id FROM songs WHERE song_tempo between \""+num_tempo+"\"" + "and \""+num_tempo2+"\"" + "AND song_loudness<\""+num_loudness+"\"" +
                             notNull+")";
                 }else {
-                    num_tempo = numTempo[0] - (float) priority.get(prioTempo);
+                    num_tempo = numTempo[0] - (double) priority.get(prioTempo);
                     q = mapGenre +
                             "(Select song_artist_id FROM songs WHERE song_tempo>\"" + num_tempo + "\"" + "AND song_loudness<\"" + num_loudness + "\"" +
                             notNull + ")";
@@ -137,7 +137,7 @@ public class Query {
      */
 
     public String UserInput(String genre, String loudness, String tempo,String prioGenre, String prioLoudness, String prioTempo,boolean popular){
-        HashMap<String,Float> priority = Maps.getInstance().PutInPriority(prioLoudness,prioTempo);
+        HashMap<String,Double> priority = Maps.getInstance().PutInPriority(prioLoudness,prioTempo);
         List<String> couples=new ArrayList<>();
         List<String> otherGenre=new ArrayList<>();
         if (prioGenre.equals("medium") || prioGenre.equals("low")){
